inline/move http.make_* functions
