+++ mitmproxy/proxy/layers/http/_http1.py
@@ -1,11 +1,12 @@
 import abc
+import html
 from typing import Union, Optional, Callable, Type
 
 import h11
 from h11._readers import ChunkedReader, ContentLengthReader, Http10Reader
 from h11._receivebuffer import ReceiveBuffer
 
-from mitmproxy import http
+from mitmproxy import http, version
 from mitmproxy.net.http import http1, status_codes
 from mitmproxy.proxy import commands, events, layer
 from mitmproxy.proxy.context import Connection, ConnectionState, Context
@@ -145,13 +146,13 @@ class Http1Connection(HttpConnection, metaclass=abc.ABCMeta):
                 yield from self.make_pipe()
                 return
             connection_done = (
-                    http1.expected_http_body_size(self.request, self.response) == -1
-                    or http1.connection_close(self.request.http_version, self.request.headers)
-                    or http1.connection_close(self.response.http_version, self.response.headers)
-                    # If we proxy HTTP/2 to HTTP/1, we only use upstream connections for one request.
-                    # This simplifies our connection management quite a bit as we can rely on
-                    # the proxyserver's max-connection-per-server throttling.
-                    or (self.request.is_http2 and isinstance(self, Http1Client))
+                http1.expected_http_body_size(self.request, self.response) == -1
+                or http1.connection_close(self.request.http_version, self.request.headers)
+                or http1.connection_close(self.response.http_version, self.response.headers)
+                # If we proxy HTTP/2 to HTTP/1, we only use upstream connections for one request.
+                # This simplifies our connection management quite a bit as we can rely on
+                # the proxyserver's max-connection-per-server throttling.
+                or (self.request.is_http2 and isinstance(self, Http1Client))
             )
             if connection_done:
                 yield commands.CloseConnection(self.conn)
@@ -211,7 +212,7 @@ class Http1Server(Http1Connection):
             yield from self.mark_done(response=True)
         elif isinstance(event, ResponseProtocolError):
             if not self.response:
-                resp = http.make_error_response(event.code, event.message)
+                resp = make_error_response(event.code, event.message)
                 raw = http1.assemble_response(resp)
                 yield commands.SendData(self.conn, raw)
             yield commands.CloseConnection(self.conn)
@@ -360,6 +361,37 @@ def make_body_reader(expected_size: Optional[int]) -> TBodyReader:
         return Http10Reader()
     else:
         return ContentLengthReader(expected_size)
+
+
+def make_error_response(
+    status_code: int,
+    message: str = "",
+) -> http.Response:
+    body: bytes = """
+        <html>
+            <head>
+                <title>{status_code} {reason}</title>
+            </head>
+            <body>
+                <h1>{status_code} {reason}</h1>
+                <p>{message}</p>
+            </body>
+        </html>
+    """.strip().format(
+        status_code=status_code,
+        reason=http.status_codes.RESPONSES.get(status_code, "Unknown"),
+        message=html.escape(message),
+    ).encode("utf8", "replace")
+
+    return http.Response.make(
+        status_code,
+        body,
+        http.Headers(
+            Server=version.MITMPROXY,
+            Connection="close",
+            Content_Type="text/html",
+        )
+    )
 
 
 __all__ = [
